+++ package.json
@@ -25,7 +25,7 @@
   ],
   "scripts": {
     "pre-push": "yarn lint && yarn build && yarn test && yarn doc",
-    "update-scalecube": "lerna clean --yes && rimraf node_modules/@scalecube && yarn install && yarn bootstrap",
+    "update-scalecube": "lerna clean --yes && rimraf node_modules && yarn install && yarn bootstrap",
     "start": "http-server -a localhost -p 8000",
     "start-node-example": "lerna run start-node-example",
     "bootstrap": "lerna bootstrap --npm-client=yarn",
