+++ packages/api/src/cluster/ClusterEvent.ts
@@ -1,0 +1,23 @@
+import { MemberEventType } from './index';
+
+/**
+ * @interface ClusterEvent
+ * Notification from Cluster to the subscriber
+ */
+export interface ClusterEvent {
+  /**
+   * @property
+   * 'ADDED' | 'REMOVED' | 'INIT' | 'CLOSE'
+   */
+  type: MemberEventType;
+  /**
+   * @property
+   * data that pass between members
+   */
+  items: any[];
+  /**
+   * @property
+   * from which member the information arrived
+   */
+  from: string;
+}
