+++ packages/api/src/cluster/MemberMap.ts
@@ -1,0 +1,12 @@
+export interface MembersMap {
+  membersState: MembersData;
+  membersPort: MembersPort;
+}
+
+export interface MembersData {
+  [member: string]: any;
+}
+
+export interface MembersPort {
+  [member: string]: MessagePort;
+}
