+++ packages/api/src/cluster/index.ts
@@ -1,3 +1,19 @@
 import { ConnectWorkerEvent, ConnectWorkerEventType } from './ConnectWorkerEvent';
+import { MemberEventType, MembershipEvent } from './MemberShip';
+import { ClusterEvent } from './ClusterEvent';
+import { Cluster, JoinCluster, ClusterOptions } from './Cluster';
+import { MembersMap, MembersPort, MembersData } from './MemberMap';
 
-export { ConnectWorkerEvent, ConnectWorkerEventType };
+export {
+  ConnectWorkerEvent,
+  ConnectWorkerEventType,
+  MemberEventType,
+  MembershipEvent,
+  ClusterEvent,
+  Cluster,
+  JoinCluster,
+  ClusterOptions,
+  MembersMap,
+  MembersPort,
+  MembersData,
+};
