+++ packages/cluster-browser/.gitignore
@@ -1,0 +1,7 @@
+.cache
+node_modules
+
+lib/
+es/
+
+report.html
