+++ packages/cluster-browser/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  transform: {
+    '.(ts|tsx)': 'ts-jest',
+  },
+  testRegex: '(\\.|/)spec\\.ts$',
+  testPathIgnorePatterns: ['<rootDir>/es/', '<rootDir>/lib/', '<rootDir>/node_modules/'],
+  moduleFileExtensions: ['ts', 'tsx', 'js'],
+  moduleDirectories: ['node_modules', 'app/src'],
+};
