+++ packages/cluster-browser/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@scalecube/api",
-  "version": "0.0.2-alpha.7",
+  "name": "@scalecube/cluster-browser",
+  "version": "0.0.1",
   "private": false,
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -12,21 +12,27 @@
   "license": "MIT",
   "scripts": {
     "clean": "rimraf node_modules && rimraf .cache && rimraf lib && rimraf es",
-    "build": "rimraf lib && yarn build-rollup",
+    "build": "yarn build-rollup",
     "build-rollup": "rollup -c",
     "lint": "tslint '{src,tests}/**/*.{ts,tsx}' --fix",
     "prettier": "prettier --write '{src,tests}/**/*.{ts,tsx}'",
-    "coverage": "yarn test-dom-env --coverage --collectCoverageFrom='src/**/*.**' --collectCoverageFrom='!tests/**/*.**'",
-    "doc": "typedoc ./src --out ./doc --mode file --name 'Scalecube API' --hideGenerator --readme ./README.md"
+    "test": ""
   },
   "author": "Scalecube (https://github.com/scalecube/scalecube-js)",
   "devDependencies": {
+    "@types/jest": "^23.3.13",
     "@types/node": "10.1.4",
+    "jest": "^24.6.0",
     "rollup": "^1.14.6",
     "rollup-plugin-filesize": "^6.1.1",
     "rollup-plugin-typescript2": "^0.21.1",
+    "rollup-plugin-visualizer": "^2.2.0",
+    "ts-jest": "^23.10.5",
     "tslint": "^5.11.0",
-    "typedoc": "^0.14.2",
     "typescript": "^3.2.4"
+  },
+  "dependencies": {
+    "@scalecube/api": "^0.0.2-alpha.7",
+    "@scalecube/utils": "^0.0.2-alpha.7"
   }
 }
