+++ packages/cluster-browser/rollup.config.js
@@ -1,0 +1,29 @@
+import visualizer from 'rollup-plugin-visualizer';
+import typescript from 'rollup-plugin-typescript2';
+import filesize from 'rollup-plugin-filesize';
+import pkg from './package.json';
+
+export default {
+  input: 'src/index.ts',
+  output: [
+    {
+      file: pkg.main,
+      format: 'cjs',
+    },
+    {
+      file: pkg.module,
+      format: 'es',
+    },
+  ],
+  external: [...Object.keys(pkg.dependencies || {}), ...Object.keys(pkg.peerDependencies || {})],
+  plugins: [
+    typescript({
+      typescript: require('typescript'),
+      clean: true,
+    }),
+    visualizer({
+      filename: 'report.html',
+    }),
+    filesize(),
+  ],
+};
