+++ packages/cluster-browser/src/Cluster/JoinCluster.ts
@@ -1,26 +1,14 @@
 import { ReplaySubject } from 'rxjs';
-import { Address } from '@scalecube/api';
+import { ClusterApi } from '@scalecube/api';
 import { getFullAddress } from '@scalecube/utils';
-import { Cluster, MemberEventType, MembersMap } from '../../helpers/types';
 import { server } from './Server';
 import { client } from './Client';
-import { setLocalAddress, utils } from './utils';
+import { createMember } from './Member';
 
-interface JoinCluster {
-  address: Address;
-  seedAddress?: Address;
-  itemsToPublish: any;
-  transport: any;
-  retry?: {
-    timeout: number;
-  };
-  debug?: boolean;
-}
-
-export const joinCluster = (options: JoinCluster): Cluster => {
+export const joinCluster: ClusterApi.JoinCluster = (options: ClusterApi.ClusterOptions) => {
   const { address, seedAddress, itemsToPublish, transport, retry, debug } = options;
   const { port1, port2 } = new MessageChannel();
-  const membersStatus: MembersMap = {
+  const membersStatus: ClusterApi.MembersMap = {
     membersPort: {},
     membersState: {},
   };
@@ -28,13 +16,11 @@ export const joinCluster = (options: JoinCluster): Cluster => {
   const delayedActions: any[] = [];
   let isConnected = !seedAddress;
 
-  const rSubjectMembers = new ReplaySubject<ClusterEvent>(1);
-  const whoAmI = getFullAddress(address);
-  setLocalAddress(whoAmI);
+  const rSubjectMembers = new ReplaySubject<ClusterApi.ClusterEvent>(1);
 
   let clientPort: any;
 
-  const { updateConnectedMember, getMembershipEvent } = utils(whoAmI, membersStatus);
+  const { updateConnectedMember, getMembershipEvent, whoAmI } = createMember(address, membersStatus);
 
   const serverPort = server({
     whoAmI,
@@ -91,11 +77,5 @@ export const joinCluster = (options: JoinCluster): Cluster => {
         }
       });
     },
-  }) as Cluster;
+  } as ClusterApi.Cluster);
 };
-
-export interface ClusterEvent {
-  type: MemberEventType;
-  items: [];
-  from: string;
-}
