+++ packages/cluster-browser/src/helpers/constants.ts
@@ -1,0 +1,7 @@
+export const MEMBERSHIP_EVENT = 'membershipEvent';
+export const MEMBERSHIP_EVENT_INIT_SERVER = 'membershipEventInitServer';
+export const MEMBERSHIP_EVENT_INIT_CLIENT = 'membershipEventInitClient';
+export const MESSAGE = 'message';
+export const ADDED = 'ADDED';
+export const REMOVED = 'REMOVED';
+export const INIT = 'INIT';
