+++ packages/cluster-browser/src/helpers/utils.ts
@@ -1,0 +1,43 @@
+import { ClusterApi } from '@scalecube/api';
+import { MEMBERSHIP_EVENT } from './constants';
+
+export const getKeysAsArray = (obj: {}) => (obj && Object.keys(obj)) || [];
+
+let localAddress: any[] = [];
+
+export const setLocalAddress = (address: string) => {
+  localAddress = [...localAddress, address];
+  return localAddress;
+};
+
+export const saveToLogs = (msg: string, extra: { [key: string]: any }, debug?: boolean) => {
+  // tslint:disable
+  debug && extra && extra.membersState && console.log(msg, 'membersState: ', extra.membersState);
+  debug && extra && extra.membersPort && console.log(msg, 'membersPort: ', getKeysAsArray(extra.membersPort));
+  // tslint:enable
+};
+
+export const genericPostMessage = (data: any, transfer?: any[]) => {
+  try {
+    // @ts-ignore
+    if (typeof WorkerGlobalScope !== 'undefined' && self instanceof WorkerGlobalScope) {
+      if (data.detail && data.detail.to && localAddress.indexOf(data.detail.to) > -1) {
+        const event = new MessageEvent('message', {
+          data,
+          ports: transfer ? transfer : undefined,
+        });
+        dispatchEvent(event);
+      } else {
+        // @ts-ignore
+        postMessage(data, transfer ? transfer : undefined);
+      }
+    } else {
+      if (data.type === 'ConnectWorkerEvent') {
+        return;
+      }
+      postMessage(data, '*', transfer ? transfer : undefined);
+    }
+  } catch (e) {
+    console.error('Unable to post message ', e);
+  }
+};
