+++ packages/cluster-browser/src/index.ts
@@ -1,0 +1,3 @@
+import { joinCluster } from './Cluster/JoinCluster';
+
+export { joinCluster };
