+++ packages/scalecube-discovery/package.json
@@ -42,6 +42,7 @@
   "dependencies": {
     "@scalecube/api": "^0.0.2-alpha.7",
     "@scalecube/utils": "^0.0.2-alpha.7",
+    "@scalecube/cluster-browser": "^0.0.1",
     "rxjs": "^6.4.0"
   }
 }
